README: updated wording
As of today, the Spring project does not plan further Spring Boot 3 minor
releases, meaning there will be no Spring Boot 3.6, etc.
Spring Boot 4 is planned for Nov 2025, but we cannot at this point
guarantee that existing Pre-Liquibase will be compatible with it. It may
or may not require a new release of Pre-Liquibase when the time comes.